Resolve #1067 -- [PacketCmd -> GamePacketID] PacketReader (#1066)
* PacketReader:
* Changed all usages of PacketCmd with GamePacketID & LoadScreenPacketID.
* PacketHeader:
* Cmd variable changed to GamePacketID.
* PacketHandlerManager:
* Separated convertorTable variable into game and loadScreen to account for PacketCmd replaced with GamePacketID.
* PacketType:
* Replaced PacketCmd with GamePacketID.
* Added LoadScreenPacketID as an option.

+++ GameServerLib/Chatbox/Commands/MobsCommand.cs
@@ -41,7 +41,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
                 .Where(xx => xx.Value.Team == team.ToTeamId())
                 .Where(xx => xx.Value is Minion || xx.Value is Monster);
 
-            var client = _playerManager.GetPeerInfo((ulong)userId);
+            var client = _playerManager.GetPeerInfo(userId);
             foreach (var unit in units)
             {
                  _game.PacketNotifier.NotifyPing(client, new Vector2(unit.Value.Position.X, unit.Value.Position.Y), 0, Pings.PING_DANGER);
