+++ GameServerLib/Packets/PacketHandlers/HandleHeartBeat.cs
@@ -24,7 +24,7 @@ namespace LeagueSandbox.GameServer.Packets.PacketHandlers
             var diff = req.AckTime - req.ReceiveTime;
             if (req.ReceiveTime > req.AckTime)
             {
-                var peerInfo = _playerManager.GetPeerInfo((ulong)userId);
+                var peerInfo = _playerManager.GetPeerInfo(userId);
                 var msg = $"Player {peerInfo.PlayerId} sent an invalid heartbeat - Timestamp error (diff: {diff})";
                 _logger.Warn(msg);
             }
